Controller server implementation
Summary:
Fill out implementation for the controller's basic RegisterAgent and ProcessData functions.
RegisterAgent just generates a uuid for the agent, and stores the agent's host config in memory in a map.
ProcessData accepts streaming data from the agent and writes the data into a file under the agent's uuid.

Test Plan: Unit test for RegisterAgent and ProcessData.

Reviewers: zasgar, kgandhi

Reviewed By: zasgar

Maniphest Tasks: T3

Differential Revision: https://phab.corp.pixielabs.ai/D29

GitOrigin-RevId: dd69f432d0227e02258daff589c2080f7319abce