PL-38 Fix go:generate checks in FileChecker arc unit test
Summary: go:generate be given any command to execute. Update the FileChecker unit test so that it doesn't fail when a new command is given. Also add in a case for the mockgen command

Test Plan: Tried it out on the go-bindata and mockgen cases. Also tried a case with a random command

Reviewers: zasgar, kgandhi

Reviewed By: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D69

GitOrigin-RevId: 219633ebfe1c28a0cc375b0939337383abcc33d6