PL-79: Load generator deployment on separate node for GKE
Summary:
Updating and moving loadgenerator.yaml to kubernetes_manifests.
Documentation on how to taint a K8S node so that only services/deployments/pods
that tolerate the taint are scheduled on it. Also, adds documentation on how to
label a node so that the deployment can target a particular node.

Test Plan:
Tested this on a GKE cluster and observed that the loadgenerator for
hipstershop only runs on the intended node and also that no other services are
running on it.

Reviewers: #engineering, michelle

Reviewed By: #engineering, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D109

GitOrigin-RevId: 1b2b2f153c882b8258b7045bf42d5dbd2ac2697a