PL-87 Move loadgen service code to our repo
Summary: Before, you should have to create the docker image for the loadgen service inside of the hipstershop repo. Since the docker image requires the locust wrapper, templates, and configs, you would have had to copy all of those files over to the hipstershop repo. Instead, build everything for the loadgen service within our repo. This process should be generalizable to other applications as well, such as superset

Test Plan: Went through the steps in the readme -- built the docker image, pushed the image, deployed to GKE, and checked the logs to make sure the loadgen service was generating requests.

Reviewers: zasgar, philkuz, kgandhi

Reviewed By: zasgar, kgandhi

Differential Revision: https://phab.corp.pixielabs.ai/D117

GitOrigin-RevId: 269b08115ba1e88873562608ab5f9e51ac14a5e4