PL-80: Collect application performance data on GKE with bcc
Summary:
Using a compiled docker image of bcc (Ubuntu xenial) and adding a
shell script to collect application performace data. The command for collecting
data can be modified in the daemonset yaml using the options for the shell
script.

TODO: Subsequent diffs will tie data collection to the load generation for the sample
applications that we have.

Test Plan: Deployed on GKE cluster and collected data for sample application

Reviewers: #engineering, michelle

Reviewed By: #engineering, michelle

Subscribers: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D126

GitOrigin-RevId: d2aa62b86e5bd59c1bb6466cbea279f4e70bfe5e