PL-86: Adding ability to write out performance data collected using bcc to a gcs bucket
Summary:
Each node in a cluster will now write its performance data to a gcs bucket. The output also has an env.txt
file that contains information about the node and the pod to identify the host and pod, respectively.
1. Modified dockerfile to install gcsfuse. This tool is used to mount gcs buckets for reading/writing
2. Updated the deployment yaml files to add the necessary commands to mount the pl-tmp gcs bucket
3. Modified the bcc commands script to write out data to a folder in a mounted gcs bucket. The output
   folder name is time stamped to avoid overwriting.

Test Plan: Tested this on a GKE cluster and was able to observe data being written to gcs bucket.

Reviewers: #engineering, michelle

Reviewed By: #engineering, michelle

Subscribers: zasgar, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D132

GitOrigin-RevId: 603d44964e2e42e31561ed17336d45ef007cda12