PL-247: Agent config proto for pub/sub control interface
Summary:
Defining the proto for configuring an agent. The data collector
publishes the set of data metrics available to it along with metadata on the
node and sends this as a publish message to an agent. The agent in response
subscribes to a subset of the publish message and assigns IDs to InfoClass and
sends back a subscribe message.

Both the data collector and agent are then expected to set up arrow tables based
on the subscribe message.

Test Plan: N/A.

Reviewers: #engineering, zasgar, oazizi

Reviewed By: #engineering, zasgar, oazizi

Subscribers: philkuz, michelle

Differential Revision: https://phab.corp.pixielabs.ai/D221

GitOrigin-RevId: 8977613bffdd821c864c76f27a87ad1c7a42553d