PL-328 Rename old compilerState -> planState, add compilerState
Summary:
we had an existing class called CompilerState, which contains the udf/uda registry. We don't want to pass this info around for our actual compilerState, so renamed this class to planState.
Added a new compilerState class which tracks the registry lookup and schema.

Test Plan: existing tests should pass

Reviewers: #engineering, zasgar, philkuz

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D282

GitOrigin-RevId: e13327eb43316e458fc161cf81dc53f4c774b831