PL-241: Creating a data source connector class for proc stat
Summary:
Add a ProcStatConnector to test out the path from a source connector
that reads data from a file and is able to populate a data table based on a
common InfoClassSchema. Note that the source connector itself is temporary for
M2. But this allows us to test a bunch of stuff.

Test Plan:
Added a unit test to get data from a source and populate a data
table.

Reviewers: #engineering, oazizi, zasgar

Reviewed By: #engineering, oazizi, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D279

GitOrigin-RevId: ceddc2e7d01f1a0da22c08604971648c07221f30