PL-241 Change source registry to store function to create source connector
Summary:
The source registry used to hold the source connector objects. Instead
it would be better for the registry to hold the function to create a source
connector and let the data collector create the source connector and own the
source connector. This way the registry has all the information regarding a
source and the data collector can create sources as needed. In the future we
could apply filters in the data collector based on type, etc. to create only
sources that it needs.

Test Plan: Updated unit test

Reviewers: #engineering, oazizi, zasgar

Reviewed By: #engineering, oazizi, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D288

GitOrigin-RevId: f297da3b7fbe23b6a2c437ec51503e6245e7f120