PL-184 IRGraphWalker class
Summary:
add an IRWalker class which walks through the IR nodes in an IR operator graph. This class is heavily based on what we already have for the PlanFragmentWalker and ScalarExpressionWalker.
We'll use this walker in the compiler, which will use it to walk through all of the nodes and convert those nodes to their protobufs.

Test Plan: added unit test

Reviewers: philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D301

GitOrigin-RevId: 8f22bce8278b7b44bdf767f1d7f06d7acb6fdfd3