PL-327 Carnot Class: Wrapper for Compiler & Execution Engine
Summary:
Added the Carnot class, which brings the compiler and execution engine together. The class keeps track of necessary state. When a query is made, it get the logical plan from the compiler, and then executes the logical plan.

Also fixed some bugs in the planGraph class.

Test Plan: added unit test for planWalker. we should write component tests once the compiler is ready.

Reviewers: #engineering, zasgar, philkuz, kgandhi

Reviewed By: #engineering, zasgar, philkuz, kgandhi

Subscribers: kgandhi

Differential Revision: https://phab.corp.pixielabs.ai/D287

GitOrigin-RevId: 8a5f5a3bb0be7ea6712b6833b6e4d26ad22ea4ca