PL-341: Adding availability function to determine whether Stirling can use a SourceConnector
Summary:
For the registry, we use a virtual function in the Source Connectors to
determine whether a system can support that Source Connector. Based on whether
it is available, Stirling can add it as a source from which to gather data. For
any other data source dependencies, we plan to add a config_setting to exclude
headers and the appropriate code to be included in the binary.

Test Plan:
Manually tested fake proc stat data with the Stirling wrapper to
ensure that the source was available

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D308

GitOrigin-RevId: b2a3f50a6643ac1bafb5bf9df55d50ef309651fe