PL-326 Collapse Range IR into MemorySourceIR
Summary:
RangeIR is the only operatorIR which does not have a 1:1 mapping with a node in the logical plan. Before converting our IR to the logical plan, we should merge the rangeIR with the memorySourceIR.
This diff collapses the range IR, but doesn't yet convert the string ("-2h", "-10m", etc) into the correct ms.

Test Plan: added unit test

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D307

GitOrigin-RevId: 90941617a4614b3f2a0663759221e67ba8722a0b