PL-364: Adding a way to discover publish proto in Stirling
Summary:
Adding a function in Stirling to get a publish proto message.
Updating Stirling to modify info schemas and data tables based on subscriptions.
Modified proto to have only two states for an element (subscribed and not subscribed), since
ownership of tables belongs to the agent and not Stirling.

Test Plan: Added a new component test.

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D333

GitOrigin-RevId: f49196142bc82c5e4b71794c23843ae17ad2cbca