Fix conflicts with GRPC CC and GO rules
Summary:
After spending way too much time tracking down incompatibilities with the rules,
here is the fix I came up with.

The problem is that the GRPC CC rules have an issue dealing with both the GOGO proto and the default proto rules. The GO rules
can't rely on the GRPC rules so there an inherent issue.

One of things to consider is if we should just have a pl_proto_libraries that generates the CC, go files,
instead of calling them out separately.

Test Plan: Build

Reviewers: michelle, oazizi, #engineering, kgandhi

Reviewed By: oazizi, #engineering, kgandhi

Subscribers: kgandhi

Differential Revision: https://phab.corp.pixielabs.ai/D346

GitOrigin-RevId: f011a14c260618122f8200d8d549867060f7afe9