PL-380 Make MemorySourceNode work with range
Summary: if start_time or stop_time are set in the MemorySource op, it assumes that there is an implicit time_ column in the relation. it finds the correct batch and index for the start and end times using binary search. when calling GenerateNext, it will return only the rows in those start and end times.

Test Plan: added unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D372

GitOrigin-RevId: 684d5914a5bce49e9ca652da4e04c9f997cf6ff5