Stirling clean up: Subscribe to an info class instead of elements
Summary:
We used to add subscription per element in an info class. But the more
useful and common usecase is to subscribe to all the elements in an infoclass.
So adding a subscribe bool to the proto and change default for every element in
an infoclass to be subscribed. Now an agent can just subscribe to an infoclass
and all its elements will be included. Simplifies the interface a little bit.

Test Plan: existing tests

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D425

GitOrigin-RevId: c6c84ddcae97fc0cd1c96f409027ec488b14bd32