Add vizier service handlers
Summary:
This adds the Vizier service handler that is responsible for holding agents and issueing queries to them.
This is M2 complete, but needs a lot more work (ie. to be more robust against agent disconnects, etc.). Will work on that
after 1.7/2.0 code is checked in.

Test Plan: Coming soon...

Reviewers: michelle, oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D476

GitOrigin-RevId: 12188a8c3a0103da787e012a21145a5266810bde