PL-470: Cleaning up source name as metadata in info class manager
Summary:
We consolidated names for the source connectors and info class
managers. There is a 1:1 relationship between source connectors and info class
managers. There is no need for source name to be part of the metadata since this
is the same as the info class manager name.

Test Plan: existing tests

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D525

GitOrigin-RevId: 3d472ec21bb8de0aed60d098ef940a9ce55cabd7