http_trace_connector: gunzip support, more filtering, and some clean-up
Summary:
This add gunzip support, and restructures the position of the filtering function a bit.

Some other stylistic and clean-up changes too. @yzhao: hope I'm not changing too much on you.

Also, I moved most of the functions to be private static class members.

Test Plan: Ran it on GKE.

Reviewers: yzhao, #engineering

Reviewed By: yzhao

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D654

GitOrigin-RevId: ae9ffa9336022e534165ab8fc178a95038f06f95