Use TGID with FD to avoid potential aliasing
Summary:
File descriptors are per user-space PID (kernel-space TGID),
so to uniquely identify events in the perf buffer, we have
to tag them with {TGID, FD}.

This is part 1, where we fix this in the BPF program.
A separate diff will fix this in the user-space program.

Test Plan: Tested on the dev cluster.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D680

GitOrigin-RevId: bba2737ab8bb72d3ce95ceba07130871668fcc1b