Add simple chunk parsing to HTTP trace.
Summary:
It now only detects a single chunk. I.e., if the message body has multiple chunks, it will stop at
the first chunk.

In practice, it seems silly to write multiple chunks in write() syscall, instead of combining them
into single chunk. I.e., given:
7\r\n
1234567\r\n
3\r\n
123\r\n
It's obvious that they should be combined into a single chunk:
10\r\n
1234567123\r\n

The next step is to handle this case correctly.

On the other hand, if the message body is not even a complete chunk, the parsing code works just
fine.

Test Plan: unit test

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Subscribers: #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D675

GitOrigin-RevId: 692e0dcfa8992d27c94ce7696e25d71bb663dd2d