(PL-525) Adding error reporter for any nodes that don't have a connection to a sink in the compiler.
Summary:
If nodes don't connect to a sink, then segfaults occurred in the executor and the error was opaque to the compiler.
Quick fix is to disallow nodes that don't touch a sink.

Test Plan: Added a test to compiler and carnot with nodes that don't touch Result. In CarnotTest, pre this commit the executeQuery call would segfault. In CompilerTest the query would compile, which now will fail instead.

Reviewers: michelle, zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D685

GitOrigin-RevId: a4c6d04deb6c8c85abffef39045ef3de218be559