Add clang-tidy to build, fix some related breaks
Summary:
This adds clang-tidy to the build and fixes a related compilation error.
This does not add clang-tidy to linting or auto-fixing. It will only stop
for build breaks related to clang-tidy and not style enforement warnings.
Changes for checking style and errors at lint time are coming.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D712

GitOrigin-RevId: 38fa4aa28b4610159ed5cf2753ab3f00d4ef4288