HTTP probe for send
Summary:
Adds a probe to send in addition to write.

This diff could be even simpler, if we didn't want to differentiate in user-space between the two cases.

However, I wanted to test whether the send() case was being triggered, so there's actually two trace functions with a common body.

Test Plan: Nothing in sock-shop is triggering the send() version yet...so not well-tested. :(

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D711

GitOrigin-RevId: 1ca62362d8c5662f57abe87627802c848b530d52