(PL-443) Summarizing common operators into a simple, templated call.
Summary:
The current situation for adding new ops is fairly drawn out and time consuming. You typically have to edit several different sources to make everything work. The goal of this change is to make adding an op as simple as editing only two places in the code.

So far this diff implements Agg op and Map op. In a later diff, all except Range and RangeAggOp can be implemented in this way.

Test Plan: tests had to be adjusted, but existing tests pass after certain architectural changes are accounted for.

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D732

GitOrigin-RevId: ad09b8e448e1305045d3347dc6c9d0923dcffa84