FS Watcher for Inotify Event Updates
Summary:
Adding an fs watcher for any generic filesystem. It uses inotify to
indicate events that are added by a caller using AddWatch for a given filesystem
path and appropriate flags. A caller can also remove existing watchers. And
reset the watcher in case of an overflow or for other reasons.

Test Plan: Added unit tests.

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D734

GitOrigin-RevId: 86aaf68f1f856b819d605f8415b760e5860e50b6