(PL-518) Adding compiler error context and better line/col reporting.
Summary:
Adding proper formatting for line/col errors in the Status. Should make it much easier
to float this information up to the user.

Isolated everything into a CompilerErrorBuilder for ease of use.

Test Plan: Tests pass, added another test for making sure number of errors are correct. A future commit should contain tests that properly check the line and column of errors that are produced.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D746

GitOrigin-RevId: 4071d7e5a75d1035f3ad2cd366ccf869630b7c19