Update querybroker pb
Summary:
As an intermediate step, before we implement compute nodes and result merging, have the query broker expect to receive results individually from each agent (like we do in the demo).
Update the protobuf so that agents have a way of sending these results to the query broker.

Test Plan: n/a

Reviewers: zasgar, philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D753

GitOrigin-RevId: 4cf7b74a3c561b4782471bbc74c68a4731971b2e