Implement QueryExecutor, to be used in QueryBroker
Summary:
The QueryExecutor (pulled out from query_broker/controllers/server.go in the ts-demo branch) is the module responsible for sending out the query fragments to the agents, and for waiting for their responses.
For now, the QueryExecutor just broadcasts the same query to all agents and expects to hear back from all agents.

Test Plan: unit test

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D755

GitOrigin-RevId: 3af81686e2dc24007c0caedc49aeed031fb3defb