Adding protobuf message for RelationMap.
Summary:
Need to add a message that is a condensed form of the RelationMap. I threw it into the compiler status but I think this might not be the most appropriate place for it.
I also wondered whether we had a schema message, but I couldn't find any so this will do.

Test Plan: Passes previous tests, should work with this.

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D786

GitOrigin-RevId: bcbd2bcca52db2776d3889ec3ff81d1eb2e1f4e2