PL-555 Updating plan proto and populating plan proto with more info
Summary:
This is the first in a few commits to split up compiler and executor. Previously, we couldn't separately pass a LogicalPlan into Carnot.
adding arg_types support for scalarfunc and aggregateexpression in plan

Pushing arg types into the FuncIR in the Compiler

Adding changes to compiler tests after proto updates

Test Plan: Updated compiler tests to include the new proto field in the resulting proto.

Reviewers: michelle, zasgar, kgandhi, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D781

GitOrigin-RevId: 68abb5bde9a5ef92d10141c318562959df555ce3