ShareLaTex Application for demos (Initial k8s manifests)
Summary:
Adding kubernetes manifest files for deploying a fully working version
of ShareLatex. Added a README with a summary on deployment and next steps. Currently,
all microservices are deployed as a single service. Next step is to separate them out.

Test Plan:
Deployed on GKE. Tested that multiple users can collaborate on
a project and chat. Verified that most services work.

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D796

GitOrigin-RevId: 0bacf976dc5f94ee758fde6882d18add78bb47a2