PL-555 Refactor carnot to accept plan instead of string query.
Summary:
part 2 of several commits to allow logical plan to be passedf directly to carnot.

Would appreciate any suggestions on how to test this feature out and make it more robust to errors in the Logical Plan.

Another question to answer is whether we should keep the udf ids in the logical plan. I could see an issue where we writeover
one definition of a udf when we accidentally share the same id in the logical plan. Currently, the compiler shouldn't cause this problem,
but maybe it would be good to act proactively in case the issue does arise?

Test Plan: Passes carnot tests, would like suggestions on how else to test this.

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D782

GitOrigin-RevId: 5310202d5ef3b06ed543ba6c34243dcb34755999