Add HTTPParser class that can stitch events from a HTTP message with Content-Length header.
Summary:
We rely on the fact that all writing to a fie descriptor has consecutive sequence numbers. So the
missing sequence numbers indicates missing events. And we the current event's previous sequence
number to find the previous partial message, which should be the previous part of this current
event.

Cleanup:
- http_header_keys -> http_headers shorten the text.
- HTTPTraceEventType -> SocketTraceEventType: I intend to make this as the type for all
  events traced on socket connections. Please let me know what you think.

Test Plan: Unit tests.

Reviewers: oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D817

GitOrigin-RevId: 1a5033fababce92bce0ed329bdda23f497e55e64