Moving fs watcher to common
Summary:
Moving fs watcher to src/common/base since this can be a common utility.
If common/base doesn't seem like the right place, may be a new directory for common
utilities? Any other suggestions?

Test Plan:
Exisiting tests. Updated the inotify event test. The reason for the flappiness
earlier was that inotify would report multiple events for the same file change with the
same event mask value. Most of the time it only reports a single event. I have added a TODO
to uniquify events in the event queue.

Reviewers: zasgar, oazizi, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D812

GitOrigin-RevId: a1ade222a0f70a91d7a8a5878a7e92f4a8148e74