Stop using manually named indexes for DataTableSchema
Summary:
Putting this out there for discussion.

I don't think the manually named indexes into the DataTableSchema solved any true maintainability problems,
because we still had to make sure there was a match between the DataTableSchema and the named indexes.

Instead, this uses the idx++ method, and adds a check to make sure all columns are populated.

Also a KeyIndex() function is added to DataTableSchema to enable more robust named access to fields.
KeyIndex() is still not used in TransferData() (and helpers) to ensure performance.

What do folks think?

Test Plan: Existing tests.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D840

GitOrigin-RevId: cb857e7450c6683683bb47d1d083ef05f821ab1b