Trace both flavors of accept()
Summary:
Reason we were missing out on some of the JSON in sock-shop is because we originally only tracked
connections for which we saw an accept().

A separate change tried to infer the connection type even when no accept was seen, and
this enabled us to capture the missing services in sock-shop.

The real root cause, however, was that there are two accept syscalls, and we were only tracking
the accept4() syscall, and not the accept() syscall.

This fixes that, which also gives us the destination IP address for these connections.

Test Plan: stirliing_wrapper

Reviewers: zasgar, yzhao, kgandhi, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D878

GitOrigin-RevId: 9bee6a3af6eb846769c3b7c513ced0e3586000f5