Add K8s MetadataHandler
Summary:
The MetadataHandler listens to a channel for any k8s object updates. Each object type will have a separate handler function, which converts the object to a protobuf and calls the corresponding update method in the MetadataStore.
Currently, since we're still waiting for D881 to land, this is only for the "Endpoints" object so far.
MetadataStore interface to be implemented after, if the idea looks okay.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D890

GitOrigin-RevId: 0e2eb35d34c473e774d449ca3b987a449d9652b0