proc_parser bug-fix: pointers to populate must be determined non-statically
Summary:
This was caught by some new tests that will be submitted in a separate diff.

Since the result struct is being passed in as argument, we can't populate the
addresses into the map using a static variable assignment.

Instead, we keep the map as a static member, but populate its addresses every call.

Alternative is to do away with the static-ness altogether.

Test Plan: A new test is coming in a separate diff. For now, just check that everything is still functional.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D942

GitOrigin-RevId: 8ccbe196f6dffebffed409a0240501fb160f80ff