Add etcd queue
Summary:
the metadata service needs to keep track of all updates it should send to each agent. we will have a queue of updates for each agent.
etcd is simply a key-value store, but the go client has a Queue implementation which handles logic for de/enqueuing.
unfortunately, their implementation of dequeue doesn't do what we want: calling dequeue will block until there's an item on the queue.
luckily their implementation of queue is simple enough that we can copy over most of the file with minor modifications to do what we need.

Test Plan: added unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D973

GitOrigin-RevId: 4adb219e38bdab74dadee6f29f2922f27abfd52a