Add DequeueAll function to etcd queue
Summary: realized that we'll want a dequeueAll method for our etcd queue class, since we'll usually want to fetch all of the items in the queue at once. Calling Dequeue for each item in the queue is inefficient, since we'll be making a separate call to etcd each time.

Test Plan: unit test

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D977

GitOrigin-RevId: 725895b14ff4b1c53264e00335d9f39d83a0f254