Add Go script that dynamically finds user includes in bpf files and expands with the header's content.
Summary:
Add a go script that looks for #include " in socket_trace.c and pidruntime.c and replace it with the content of the header files. It also checks that the header file is included in bazel as a src, and that the header itself doesn't include any other user definied header file reference.
Add a Bazel preprocess gen rule that calls this script to expand the headers in sockettrace.c and pidruntime.c.
Changed OBJ_STRVIEW path

Test Plan: Passes stirling tests.

Reviewers: oazizi, yzhao, zasgar, #engineering

Reviewed By: oazizi, yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D962

GitOrigin-RevId: e4b380a00ba45b14ebca2fecd6ad57b5975cdfad