Allow test binary to run directly under ToT not using bazel test.
Summary:
This is useful sometimes to run test binaries outside of bazel test to
clarify if certain test behavior is relevant to bazel test environment.

For example, I saw a segment fault in bazel test but not running the
test binary directly. So this option would be help for that.

Test Plan: Manual testing.

Reviewers: oazizi, kgandhi, #engineering, zasgar

Reviewed By: #engineering, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D999

GitOrigin-RevId: 68615e5a3fc4640339390c8fc75667b861944782