Stirling: request and response matching (first version)
Summary:
This diff finds requests that correspond to responses, and matches them up in a single record.

This is just a simple version to serve as a foundation.

Assumption: No events are lost from BPF to user-space. We need to make this (and other cases)
more robust to this.

Test Plan: New test case added.

Reviewers: yzhao, zasgar, chengruizhe, #engineering

Reviewed By: yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1040

GitOrigin-RevId: 17d28a6d9baa1b552098c17fdc228c14d079fb49