Fix Agent Manager UpdateAgentState
Summary: metadata service is periodically crashing in the nightly build because of UpdateAgentState. When UpdateAgentState fetches all the agents, its also fetching the locks in etcd (because of how etcd names their lock keys). We had this issue before in GetActiveAgents, so we had to make sure to explicitly filter them out. We need to make sure to do this in UpdateAGentState as well. PUlled this out into the etcd metadata store... since we should do some cleanup and move all etcd-dependent things in agentManager to there anyway.

Test Plan: added tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1051

GitOrigin-RevId: 7b06db35f1834e004da970b730bd8af9d54e1e3e