Inferring traffic should set message type differently for send and recv
Summary:
Traffic class inference should have been based on whether it was the send or recv channel of a connection,
because this is then used to determine whether to push events into the perf buffer.

Test Plan: None yet - a better client-server system to monitor in socket_trace_bpf_test would help here.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1066

GitOrigin-RevId: f389c102c1fa440052d8d74072edca0a02216ed0