Update AgentUpdateInfo to include SchemaInfo
Summary:
we'll need to receive more schema information from the agent (start time, end time, table name) than we were originally sending in the AgentUpdateInfo message.
Added a schemaInfo message which has that information. In our data model doc, we also have some fields for cardinality and table size... but we'll add that later when we get to it

Test Plan: n/a, existing should pass

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1056

GitOrigin-RevId: 49c6f43e216155dadd5a2812e61db6a88ca81f93