Socket Tracer: Capture closed events, and cleanup fully processed connections.
Summary:
We need a way of cleaning up connections, so we don't leak memory.

This is a foundation of that infrastructure that assumes no lost events from BPF.

Lost events from BPF will be handled in the same way, but with different conditions to determine whether the connection is done.

Test Plan: New tests in socket_trace_connector_test. One with inorder events, and one where the close shows up before the last data event.

Reviewers: yzhao, zasgar, chengruizhe, #engineering

Reviewed By: yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1075

GitOrigin-RevId: c43d607c72dadee57a23060166d94793530dee13