Bug-fix: set timestamp so we don't create new connections
Summary: We were creating different generations of connections, which prevented req-response matching.

Test Plan:
socket_trace_bpf_test is the best candidate for catching this sort of bug,
but that needs to be switched to a real HTTP server-client. In theory,
a test like that would have caught this right away.

I don't want to hold back this fix in the meantime.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1105

GitOrigin-RevId: cbc244690da988a2078cd275aa542ab5d6de1b42