Add experiemental to bazelignore
Summary:
This adds experimental to .bazelignore. This will prevent experimental code from being build when
we run bazel build/test //...

Unfortunately it requires changes to bazelignore if you want to work on experimental.

I'll keep looking for a better solution, but this is the "best" I found so far.

Test Plan: N/A

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1112

GitOrigin-RevId: 7e9655c145113199472806f5fb5f6858e14c6de1