Adding start_time to PIDs in BPF code
Summary:
Getting start_time from BPF, as it will be needed for unique pids.

Currently sending this to a dedicated column of its own,
but once we support a Int128 type, will update to use a new field.

Test Plan:
New test case in socket_trace_bpf_test.cc.

Testing with time is always hard, so this uses a time window,
just to make sure number is not completely bogus.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1130

GitOrigin-RevId: 127caaa41f10ea6af9c4a3e01cd8d71d3424fd23