PL-707 Add GetPods/Endpoints to EtcdMetadataStore
Summary:
When an agent registers with the mds, the mds should send the agent all information it knows about the current k8s state (pods, endpoints). This requires a GetPods/GetEndpoints function in the etcdMetadataStore to pull that information out of etcd.
A followup diff, where the agent manager actually uses the GetPods/Endpoints methods is in progress.

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1192

GitOrigin-RevId: fb831b4e0be4db9f1bbafdd66dc70fa47f8cba0d