Adding new rule to set relations in operators.
Summary: For the datatype rule, we need to make sure that parent operations have set relations, this rule handles this for everything outside of source nodes.

Test Plan: new tests are added to try the cases within the rule itself. New tests for whenever we need to add new rules.

Reviewers: #engineering, michelle, zasgar, nserrino

Reviewed By: #engineering, michelle

Subscribers: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D1183

GitOrigin-RevId: 273d92c77a96a725363ca23324d536b783724843