PL-706 Update etcd sessions and mutexes
Summary: Realized I was using the etcd mutexes pretty incorrectly. Instead of locking on the key that I need to be updating, instead I should be locking on a "update" key. If another instance of the metadata service needs to update something, but can't get the update lock, it has to wait for the lock to be released. Also I think I was doing something weird with sessions by creating a session each time I needed a lock.

Test Plan: deployed the service and timed the amount of time it took to execute GetAgents

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1185

GitOrigin-RevId: 4a86f0501a8842e29926c5733eba63d54a69851f