Make ConsumeNext receive parent index
Summary:
When a child with multiple parents receives a row batch, it needs to
know which parent the row batch came from. This change supports that
need, which will become necessary for the zip operator.

In this way, the row batch is passed along with the index for the parent
that is passing it. So a child with parents 0 and 1 would receive 0 or 1
along with each row batch. This should be more robust to query rewriting
than using the ID of the parent node itself.

This code rests on the guarantee that children will be added to parents
in topological order. In other words, a child that has parents 0 and 1
in its input depenncies will have parent 0 add that child before parent
1 does. This seems like a safe guarantee to make but if someone sees a
flaw with it let me know.

Test Plan: These are all very simple functions, so it seems like the best coverage for this functionality is in the zip operator test. In that operator there will be tests for the cases with multiple parents.

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: michelle, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1186

GitOrigin-RevId: ca9880c483eb692489800738390428ecbde21eb7