Adding rule class that defines routines for analyzing and optimizing the compiler.
Summary:
The plan right now is to shift over the architecture to this new scheme and replace analyzer and optimizer with stuff written using these
rules.

DataTypeRule handles the setting of data types in the system.

Analyzer and Optimizer will become "RuleExecutors" that have a set of rules to run over.

Test Plan: Added a new test for the datatype.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1144

GitOrigin-RevId: 3cc25ddc6a867d333b505247ef33b59cfc318af5